FIR: move checkPropertyAccesses to FirPropertyInitializationAnalyzer
