+++ compiler/fir/checkers/src/org/jetbrains/kotlin/fir/analysis/checkers/declaration/FirMemberPropertiesChecker.kt
@@ -6,10 +6,7 @@
 package org.jetbrains.kotlin.fir.analysis.checkers.declaration
 
 import org.jetbrains.kotlin.KtFakeSourceElementKind
-import org.jetbrains.kotlin.contracts.description.canBeRevisited
 import org.jetbrains.kotlin.contracts.description.isDefinitelyVisited
-import org.jetbrains.kotlin.contracts.description.isInPlace
-import org.jetbrains.kotlin.descriptors.ClassKind
 import org.jetbrains.kotlin.descriptors.Visibilities
 import org.jetbrains.kotlin.fir.analysis.cfa.util.PropertyInitializationInfo
 import org.jetbrains.kotlin.fir.analysis.checkers.contains
@@ -18,20 +15,15 @@ import org.jetbrains.kotlin.fir.analysis.checkers.getModifierList
 import org.jetbrains.kotlin.diagnostics.DiagnosticReporter
 import org.jetbrains.kotlin.fir.analysis.diagnostics.FirErrors
 import org.jetbrains.kotlin.diagnostics.reportOn
+import org.jetbrains.kotlin.fir.analysis.cfa.checkPropertyAccesses
+import org.jetbrains.kotlin.fir.analysis.cfa.requiresInitialization
 import org.jetbrains.kotlin.fir.analysis.cfa.util.PropertyInitializationInfoData
 import org.jetbrains.kotlin.fir.declarations.*
 import org.jetbrains.kotlin.fir.declarations.impl.FirDefaultPropertyAccessor
 import org.jetbrains.kotlin.fir.declarations.utils.*
-import org.jetbrains.kotlin.fir.expressions.FirQualifiedAccess
-import org.jetbrains.kotlin.fir.expressions.FirThisReceiverExpression
-import org.jetbrains.kotlin.fir.isCatchParameter
-import org.jetbrains.kotlin.fir.references.toResolvedPropertySymbol
-import org.jetbrains.kotlin.fir.resolve.dfa.cfg.*
+import org.jetbrains.kotlin.fir.resolve.dfa.cfg.NormalPath
 import org.jetbrains.kotlin.fir.resolve.dfa.controlFlowGraph
-import org.jetbrains.kotlin.fir.symbols.FirBasedSymbol
-import org.jetbrains.kotlin.fir.symbols.SymbolInternals
 import org.jetbrains.kotlin.fir.symbols.impl.FirPropertySymbol
-import org.jetbrains.kotlin.fir.symbols.impl.FirSyntheticPropertySymbol
 import org.jetbrains.kotlin.lexer.KtTokens
 
 // See old FE's [DeclarationsChecker]
@@ -61,7 +53,7 @@ object FirMemberPropertiesChecker : FirClassChecker() {
         // TODO: this also visits non-constructor member functions...
         // TODO: merge with `FirPropertyInitializationAnalyzer` for fewer passes.
         val data = PropertyInitializationInfoData(memberPropertySymbols, symbol, graph)
-        graph.checkPropertyAccesses(memberPropertySymbols, symbol, context, reporter, data)
+        data.checkPropertyAccesses(context, reporter)
         return data.getValue(graph.exitNode)[NormalPath]
     }
 
@@ -132,99 +124,6 @@ object FirMemberPropertiesChecker : FirClassChecker() {
         ) {
             property.source?.let {
                 reporter.reportOn(it, FirErrors.REDUNDANT_OPEN_IN_INTERFACE, context)
-            }
-        }
-    }
-}
-
-val FirDeclaration.evaluatedInPlace: Boolean
-    get() = when (this) {
-        is FirAnonymousFunction -> invocationKind.isInPlace
-        is FirAnonymousObject -> classKind != ClassKind.ENUM_ENTRY
-        is FirConstructor -> true // child of class initialization graph
-        is FirFunction, is FirClass -> false
-        else -> true // property initializer, etc.
-    }
-
-@OptIn(SymbolInternals::class)
-val FirPropertySymbol.requiresInitialization: Boolean
-    get() = this !is FirSyntheticPropertySymbol && !hasInitializer && !hasExplicitBackingField &&
-            hasBackingField && fir.isCatchParameter != true
-
-fun ControlFlowGraph.checkPropertyAccesses(
-    properties: Set<FirPropertySymbol>,
-    receiver: FirBasedSymbol<*>?,
-    context: CheckerContext,
-    reporter: DiagnosticReporter,
-    data: PropertyInitializationInfoData
-) {
-    // NOTE: assert(properties.all { it.requiresInitialization })
-    // If a property has an initializer (or does not need one), then any reads are OK while any writes are OK
-    // if it's a `var` and bad if it's a `val`. `FirReassignmentAndInvisibleSetterChecker` does this without a CFG.
-    if (properties.isEmpty()) return
-
-    checkPropertyAccesses(properties, receiver, context, reporter, data, null, mutableMapOf())
-}
-
-@OptIn(SymbolInternals::class)
-private fun ControlFlowGraph.checkPropertyAccesses(
-    properties: Set<FirPropertySymbol>,
-    receiver: FirBasedSymbol<*>?,
-    context: CheckerContext,
-    reporter: DiagnosticReporter,
-    data: PropertyInitializationInfoData,
-    scope: FirDeclaration?,
-    scopes: MutableMap<FirPropertySymbol, FirDeclaration?>,
-) {
-    fun FirQualifiedAccess.hasCorrectReceiver() =
-        (dispatchReceiver as? FirThisReceiverExpression)?.calleeReference?.boundSymbol == receiver
-
-    for (node in nodes) {
-        when {
-            // TODO: `node.isUnion` - f({ x = 1 }, { x = 2 }) - which to report?
-            //  Also this is currently indistinguishable from x = 1; f({}, {}).
-
-            node is VariableDeclarationNode -> {
-                val symbol = node.fir.symbol
-                if (scope != null && receiver == null && node.fir.isVal && symbol in properties) {
-                    // It's OK to initialize this variable from a nested called-in-place function, but not from
-                    // a non-called-in-place function or a non-anonymous-object class initializer.
-                    scopes[symbol] = scope
-                }
-            }
-
-            node is VariableAssignmentNode -> {
-                val symbol = node.fir.calleeReference.toResolvedPropertySymbol() ?: continue
-                if (!symbol.fir.isVal || !node.fir.hasCorrectReceiver() || symbol !in properties) continue
-
-                if (scope != scopes[symbol]) {
-                    val error = if (receiver != null)
-                        FirErrors.CAPTURED_MEMBER_VAL_INITIALIZATION
-                    else
-                        FirErrors.CAPTURED_VAL_INITIALIZATION
-                    reporter.reportOn(node.fir.lValue.source, error, symbol, context)
-                } else if (data.getValue(node).values.any { it[symbol]?.canBeRevisited() == true }) {
-                    reporter.reportOn(node.fir.lValue.source, FirErrors.VAL_REASSIGNMENT, symbol, context)
-                }
-            }
-
-            node is QualifiedAccessNode -> {
-                val symbol = node.fir.calleeReference.toResolvedPropertySymbol() ?: continue
-                if (!symbol.isLateInit && node.fir.hasCorrectReceiver() && symbol in properties &&
-                    data.getValue(node).values.any { it[symbol]?.isDefinitelyVisited() != true }
-                ) {
-                    reporter.reportOn(node.fir.source, FirErrors.UNINITIALIZED_VARIABLE, symbol, context)
-                }
-            }
-
-            // In the class case, subgraphs of the exit node are member functions, which are considered to not
-            // be part of initialization, so any val is considered to be initialized there and the CFG is not
-            // needed. The errors on reassignments will be emitted by `FirReassignmentAndInvisibleSetterChecker`.
-            node is CFGNodeWithSubgraphs<*> && (receiver == null || node !== exitNode) -> {
-                for (subGraph in node.subGraphs) {
-                    val newScope = subGraph.declaration?.takeIf { !it.evaluatedInPlace } ?: scope
-                    subGraph.checkPropertyAccesses(properties, receiver, context, reporter, data, newScope, scopes)
-                }
             }
         }
     }
